move modern ExchangeRateText
