+++ src/components/display/ExchangeRateText.js
