+++ src/screens/Exchange/ExchangeConfirm/DetailsTable.js
@@ -23,7 +23,7 @@ import styled from 'styled-components/native';
 import t from 'translations/translate';
 
 // Components
-import ExchangeRateText from 'components/modern/ExchangeRateText';
+import ExchangeRateText from 'components/display/ExchangeRateText';
 import Table, { TableRow, TableLabel, TableAmount, TableFee } from 'components/Table';
 import { BaseText } from 'components/legacy/Typography';
 import { Spacing } from 'components/legacy/Layout';
